Add RecyclerView and CardView functionality to MainActivity for Hello World Demo
The tab layout required that I create a separate fragment for each tab.
In order to connect the MainActivity to each fragment, I had to modify the SectionsPagerAdapter to call each fragment explicitly.

I created some preliminary models to represents our two types of data, and viewModels to pass that data to the fragments.

To link the RecyclerView to the CardViews it would contain, I had to write a custom RecyclerView Adapter for each data type (Blueprints and Deployments).

We can change the way we represent the different types of data, but I figured it would be good to get granular early on.